factura no se asocia a proyecto

+++ Model/Custom/ClienteGrid.cs
@@ -12,7 +12,7 @@ namespace Model.Custom
         [Display(Name = "Nombre Completo")]
         public string ApyNom { get; set; }
         [Display(Name = "Telefono")]
-        public int Telefono { get; set; }
+        public long Telefono { get; set; }
         [Display(Name = "Email")]
         public string Mail { get; set; }
         [Display(Name = "Fecha de Nacimiento")]
